Make the statusbar show what ever is in the treemodel
Entities like textures set data in the treemodel and
we later retrive that to show in the UI. This could
be done a bit differently but we want to move towards
a more ModelViewStore model. So we can easily add new
UI elements without having to change the code that
talks rbug to the driver.